% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.r
\name{num_int}
\alias{num_int}
\title{This function performs numerical integration using monte carlo integration}
\usage{
num_int(fn, ..., a, b, N = 100, ncores = 1, cl = NULL)
}
\arguments{
\item{fn}{is a function}

\item{...}{are further arguments to be passed to fn}

\item{a}{is the lower bound}

\item{b}{is the upper bound}

\item{N}{is the sample size}

\item{ncores}{is the number of processors to call}

\item{cl}{is an object of class cluster}
}
\description{
This function performs numerical integration using monte carlo integration
}
